# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DraftCrane API
  description: |
    Backend API for DraftCrane (dc) -- a browser-based nonfiction book writing
    environment with AI assistance. Powered by Cloudflare Workers and Hono.

    ## Authentication

    Most endpoints require a valid **Clerk JWT** passed as a Bearer token in the
    `Authorization` header. Public endpoints (health check, Clerk webhook, OAuth
    callback) use their own verification mechanisms and are documented accordingly.

    ## Rate Limiting

    Standard endpoints enforce **120 requests per minute** per user (atomic
    time-window counters in D1). AI and export endpoints have tighter limits
    noted in their descriptions.

    ## Error Responses

    All errors follow a consistent shape with `error` (human-readable message),
    `code` (machine-readable enum), and `requestId` (ULID for tracing).
  version: 2.0.0
  contact:
    name: DraftCrane Engineering
  license:
    name: Proprietary
servers:
  - url: https://dc-api.automation-ab6.workers.dev
    description: Production
...
components:
  securitySchemes:
    clerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT token
  schemas:
...
    DriveConnection:
      type: object
      required:
        - id
        - email
        - connectedAt
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        connectedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Authentication required
            code: AUTH_REQUIRED
            requestId: 01HXYZ...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found
            code: NOT_FOUND
            requestId: 01HXYZ...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Title is required
            code: VALIDATION_ERROR
            requestId: 01HXYZ...
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Rate limit exceeded
            code: RATE_LIMITED
            requestId: 01HXYZ...
...
security:
  - clerkBearerAuth: []
paths:
...
  /drive/authorize:
    get:
      tags: [Drive]
      summary: Get Google OAuth authorization URL
      description: |
        Returns a Google OAuth authorization URL for connecting a Google Drive
        account. Supports `loginHint` to pre-select a specific Google account
        and `pid` to auto-link the new connection to a project after OAuth.
      operationId: getDriveAuthorizeUrl
      parameters:
        - name: loginHint
          in: query
          required: false
          schema:
            type: string
            format: email
          description: Google email to pre-select in OAuth flow
        - name: pid
          in: query
          required: false
          schema:
            type: string
          description: Project ID to auto-link after OAuth completes
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                type: object
                required:
                  - authorizationUrl
                properties:
                  authorizationUrl:
                    type: string
                    format: uri
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/callback:
    get:
      tags: [Drive]
      summary: OAuth callback
      description: |
        Google redirects here after user authorizes. Exchanges the authorization
        code for tokens, encrypts and stores them in D1, then redirects the
        browser to the frontend success or error page. Public route -- uses
        CSRF state token for verification, not Clerk JWT.
      operationId: handleDriveCallback
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: OAuth authorization code from Google
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: CSRF state token for verification
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Error code if user denied access
      responses:
        '302':
          description: Redirects to frontend success or error page
          headers:
            Location:
              schema:
                type: string
                format: uri
        '400':
          $ref: '#/components/responses/ValidationError'
  /drive/connection:
    get:
      tags: [Drive]
      summary: Get Drive connections
      description: |
        Returns all Google Drive connections for the authenticated user.
        Includes backward-compatible top-level `connected` and `email` fields
        from the first connection, plus a `connections` array for multi-account
        support. Tokens are never exposed.
      operationId: getDriveConnections
      responses:
        '200':
          description: Drive connection status
          content:
            application/json:
              schema:
                type: object
                required:
                  - connected
                  - connections
                properties:
                  connected:
                    type: boolean
                  email:
                    type: string
                    format: email
                    description: Email of the first connection (backward compat)
                  connectedAt:
                    type: string
                    format: date-time
                    description: Timestamp of the first connection (backward compat)
                  connections:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriveConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      tags: [Drive]
      summary: Disconnect all Drive accounts (legacy)
      description: |
        Disconnects all Google Drive connections for the user. Archives sources,
        soft-archives chapter-source links, clears project Drive references, and
        best-effort revokes OAuth tokens. Drive files remain untouched.
        Legacy endpoint -- prefer DELETE /drive/connection/:connectionId.
      operationId: disconnectAllDrive
      responses:
        '200':
          description: All Drive accounts disconnected
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/connection/{connectionId}:
    delete:
      tags: [Drive]
      summary: Disconnect a specific Drive account
      description: |
        Disconnects a specific Google Drive connection by ID. Cascades:
        archives sources, soft-archives chapter-source links, and best-effort
        revokes the OAuth token. Drive files remain untouched.
      operationId: disconnectDriveConnection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Drive connection disconnected
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - message
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/connection/{connectionId}/files:
    get:
      tags: [Drive]
      summary: Browse files for a specific connection
      description: |
        Browse Google Drive folders and files for a specific Drive connection.
        Supports pagination and folder navigation via folderId query param.
      operationId: browseConnectionFiles
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: folderId
          in: query
          required: false
          schema:
            type: string
            default: root
          description: Folder to browse (default "root")
        - name: pageToken
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Files in the specified location
          content:
            application/json:
              schema:
                type: object
                required:
                  - files
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriveFile'
                  nextPageToken:
                    type: string
                    nullable: true
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/connection/{connectionId}/files/{fileId}/content:
    get:
      tags: [Drive]
      summary: Get parsed file content
      description: |
        Returns the parsed text content of a Google Drive file. Supports
        Google Docs (exported as plain text), PDFs (via unpdf), and DOCX
        (via mammoth.js).
      operationId: getConnectionFileContent
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
        - name: fileId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Parsed file content
          content:
            application/json:
              schema:
                type: object
                required:
                  - content
                  - wordCount
                properties:
                  content:
                    type: string
                  wordCount:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/picker-token/{connectionId}:
    get:
      tags: [Drive]
      summary: Get picker token for a connection
      description: |
        Returns a short-lived OAuth access token for the Google Picker API,
        scoped to a specific Drive connection. Rate limited to 10 requests
        per minute.
      operationId: getPickerTokenByConnection
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Picker access token
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - expiresIn
                properties:
                  accessToken:
                    type: string
                    description: Short-lived OAuth access token
                  expiresIn:
                    type: integer
                    description: Remaining token lifetime in seconds
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Drive not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/picker-token:
    get:
      tags: [Drive]
      summary: Get picker token (legacy)
      description: |
        Returns a short-lived OAuth access token for the Google Picker API,
        using the user's first Drive connection. Legacy endpoint -- prefer
        /drive/picker-token/:connectionId for multi-account support.
        Rate limited to 10 requests per minute.
      operationId: getPickerToken
      responses:
        '200':
          description: Picker access token
          content:
            application/json:
              schema:
                type: object
                required:
                  - accessToken
                  - expiresIn
                properties:
                  accessToken:
                    type: string
                  expiresIn:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Drive not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
  /drive/folders/{folderId}/children:
    get:
      tags: [Drive]
      summary: List files in a Drive folder
      description: |
        Lists files in a Google Drive folder. Returns file metadata including
        name, MIME type, and links.
      operationId: listFolderChildren
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Files in the folder
          content:
            application/json:
              schema:
                type: object
                required:
                  - files
                properties:
                  files:
                    type: array
                    items:
                      $ref: '#/components/schemas/DriveFile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Drive not connected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/RateLimited'
        '502':
          description: Google Drive API error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...